Fixed switlint warnings
All good!
Let's hope code readability is really improved.
Replaced a couple delegates with blocks.
