+++ Timetracker/AppDelegate.swift
@@ -12,7 +12,7 @@ import AppKit
 import Preferences
 
 @NSApplicationMain
-class AppDelegate: NSObject, NSApplicationDelegate, TaskPingReceiver, NSUserNotificationCenterDelegate {
+class AppDelegate: NSObject, NSApplicationDelegate, NSUserNotificationCenterDelegate {
 
     let menu = NSMenu()
     let databaseName = Bundle.main.object(forInfoDictionaryKey: "DB_NAME") as? String ?? "default_db_name_"
@@ -141,51 +141,6 @@ class AppDelegate: NSObject, NSApplicationDelegate, TaskPingReceiver, NSUserNoti
             L.d("nothing to do, let's continue counting time")
         }
 
-    }
-
-    // MARK: - TaskPingReceiver implementation
-
-    func ping(_ interval: TimeInterval) {
-        if let button = statusItem.button {
-            let string = interval.toProperString()
-            currentTaskTime = string
-
-            var lastEvent: CFTimeInterval = 0
-            lastEvent = CGEventSource.secondsSinceLastEventType(CGEventSourceStateID.hidSystemState,
-                                                                eventType: CGEventType(rawValue: ~0)!)
-
-            if Int(lastEvent) > maxIdleSeconds {
-                if !showingIdleDialog {
-                    let now = Date()
-                    let idleDate = now - TimeInterval(maxIdleSeconds)
-                    L.d("""
-                        Showing idle dialog. \
-                        Now is \(now) and the user was idle for \(maxIdleSeconds). Started idling at \(idleDate)
-                        """)
-                    handleIdleDialog(withResponse: showIdleDialogWithIdleDate(idleDate), andIdleStart: idleDate)
-                    showingIdleDialog = false
-
-                }
-            }
-
-            button.title = builder.string()
-
-            if let menuTimer = self.menu.item(withTag: 1) {
-                menuTimer.title = string
-            }
-
-        }
-    }
-
-    func taskStopped() {
-        currentTaskTime = nil
-        _ = Timer.inOneSecond { (_) in
-            self.statusItem.button?.title = self.builder.idle
-        }
-    }
-
-    func taskStarted() {
-        statusItem.button?.title = builder.start
     }
 
     // MARK: - Core Data stack
